[Ref] Move all TypeAwareResource implementations into the util package.
